State handlers.
Now you can register a callback function for when the entity is in a
state.  This replaces the hard-coded switch statement so that I don't
have to make seperate entity classes for every type of entity.

For example:  Make a turtle entity, then register the 'crawl' handler to
advance the entity's x position and play the walk animation.