Removed the flicker! Also better debug display.
It was because when the next animation was queued up for the first time,
the x and y of the animation were not set. Thus, it rendered at an
undefined place.

Root cause was I was setting the animation x and y before deciding if
another animation should be used in the next update.  The reason the
problem only happend on the first play of an animtion was that on the
very next tick, if the animation wasn't swapped out, its x and y would
be set and remain.

I also added a more sophisticated debug display: now debug has it's own
canvas that updates every single browser animation frame.  Status for
frame and tick of game state are updated every frame.