Can change state based on animation finish.
This is needed when user input doesn't drive state change - such as
standing up from ducking:  The DUCK state is driven by user input of
holding down a key. However, when the user lifts up the down key we want
first for the entity to do a standing animation (STAND state) then
resume the IDLE state. We can know when to transition from STAND to IDLE
by polling the animation to see if it is finished.