Wrote tool to find x, y, w, h of sprites in  sheet
It was so tedious to pixel-hunt these tiny sprites off a sheet, that I
spent time writing a tool to find it - mostly it works.

The concept is to turn the image into black and white mask. The user
clicks inside the white region (the sprite) and a flood fill algo will
determine the bounding box of that sprite.

Each sprite that is clicked on gets labeled with a number, and the x, y,
w, h are printed out for that number in a list.